+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Repository/OdontologoRepository.java
@@ -9,12 +9,12 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface OdontologoRepository extends JpaRepository<Odontologo, Integer> {
+public interface OdontologoRepository extends JpaRepository<Odontologo, Long> {
     boolean existsByMatricula(@NotBlank(message = "La matrícula es obligatoria") String matricula);
 
     List<Odontologo> findByNombreContainingIgnoreCase(String nombre);
 
     List<Odontologo> findByEspecialidadContainingIgnoreCase(String especialidad);
 
-    boolean existsByMatriculaAndIdOdontologoNot(@NotBlank(message = "La matrícula es obligatoria") String matricula, Integer id);
+    boolean existsByMatriculaAndIdOdontologoNot(@NotBlank(message = "La matrícula es obligatoria") String matricula, Long id);
 }
