+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Repository/PagoRepository.java
@@ -11,7 +11,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface PagoRepository extends JpaRepository<Pago, Integer> {
+public interface PagoRepository extends JpaRepository<Pago, Long> {
     List<Pago> findByMedio(String medio);
 
     @Query("""
@@ -20,7 +20,7 @@ public interface PagoRepository extends JpaRepository<Pago, Integer> {
         JOIN Turno t ON pr.turno.idTurno = t.idTurno
         WHERE t.paciente.idPaciente = :idPaciente
     """)
-    List<Pago> findByPacienteId(@Param("idPaciente") Integer idPaciente);
+    List<Pago> findByPacienteId(@Param("idPaciente") Long idPaciente);
 
     @Query("""
         SELECT p FROM Pago p
@@ -33,7 +33,7 @@ public interface PagoRepository extends JpaRepository<Pago, Integer> {
             @Param("hasta") LocalDate hasta
     );
 
-    List<Pago> findByPracticaRealizada_Turno_Paciente_IdPaciente(Integer idPaciente);
+    List<Pago> findByPracticaRealizada_Turno_Paciente_IdPaciente(Long idPaciente);
 
     List<Pago> findByFechaPagoBetween(LocalDate fechaInicio, LocalDate fechaFin);
 }
