+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Repository/PracticaRealizadaRepository.java
@@ -11,22 +11,22 @@ import java.util.List;
 
 @Repository
 public interface PracticaRealizadaRepository extends JpaRepository<PracticaRealizada, Long> {
-    List<PracticaRealizada> findByTurno_IdTurno(Integer idTurno);
-    List<PracticaRealizada> findByPago_IdPago(Integer idPago);
+    List<PracticaRealizada> findByTurno_IdTurno(Long idTurno);
+    List<PracticaRealizada> findByPago_IdPago(Long idPago);
 
     @Query("""
         SELECT pr FROM PracticaRealizada pr
         JOIN pr.turno t
         WHERE t.paciente.idPaciente = :idPaciente
     """)
-    List<PracticaRealizada> findByPacienteId(@Param("idPaciente") Integer idPaciente);
+    List<PracticaRealizada> findByPacienteId(@Param("idPaciente") Long idPaciente);
 
     @Query("""
         SELECT pr FROM PracticaRealizada pr
         JOIN pr.turno t
         WHERE t.odontologo.idOdontologo = :idOdontologo
     """)
-    List<PracticaRealizada> findByOdontologoId(@Param("idOdontologo") Integer idOdontologo);
+    List<PracticaRealizada> findByOdontologoId(@Param("idOdontologo") Long idOdontologo);
 
     @Query("""
         SELECT pr FROM PracticaRealizada pr
@@ -40,9 +40,9 @@ public interface PracticaRealizadaRepository extends JpaRepository<PracticaReali
         WHERE t.paciente.idPaciente = :idPaciente
         AND pr.pago IS NULL
     """)
-    List<PracticaRealizada> findSinPagoByPacienteId(@Param("idPaciente") Integer idPaciente);
+    List<PracticaRealizada> findSinPagoByPacienteId(@Param("idPaciente") Long idPaciente);
 
-    List<PracticaRealizada> findByTurno_Paciente_IdPaciente(Integer idPaciente);
+    List<PracticaRealizada> findByTurno_Paciente_IdPaciente(Long idPaciente);
 
     List<PracticaRealizada> findByPagoIsNull();
 }
